Check file filename in cache instead of headers
The latest RHCOS image headers seem to omit the
Content-Disposition: inline; filename=... header so we always
download the file even when it's cached.

Instead we can download when curl -I finds the md5sum, which means
it is fully downloaded.  This won't handle the case where an image
is updated but the filename is not, but hopefully that is unlikely
given the way the image versioning is handled.