Don't compress the RHCOS image
This compression was added years ago so that IPA running in developemnt
environments with limited RAM could download the image to a tmpfs. It
added 2 to 3 minutes and at the time we cached the image (masters downloaded
it from the bootstrap VM), so it was a single hit.

Fast forward to today, we've removed the caching and now run an additional
image-cache pod on each master. So we're running this 5 times(at least 3 in series).
And the default dev env requirments is now 8G for worker nodes meaning its not needed
any longer and removing it should reduce the time taken to get a running cluster
by about 8 minutes.